+++ Block/Catalog/Product/ProductList/Banner.php
@@ -82,7 +82,7 @@ class Banner extends Template
     /**
      * Get Search results response
      *
-     * @return false|XmlElement|mixed|\SimpleXMLElement
+     * @return false|XmlElement
      */
     protected function getResponse()
     {