+++ Plugin/CatalogGraphQl/Model/Config/FilterAttributeReader.php
@@ -53,7 +53,11 @@ class FilterAttributeReader
         \Magento\CatalogGraphQl\Model\Config\FilterAttributeReader $reader,
         array $result
     ) : array {
-        $allQuestions = $this->search->getAllQuestions()->Questions->Question;
+        $allQuestions = $this->search->getAllQuestions();
+        if (empty($allQuestions)) {
+            return $result;
+        }
+
         foreach ($allQuestions as $question) {
             $attributeCode = str_replace(" ", "_", (string) $question->getAttribute('SideText'));
             if ($attributeCode) {